+++ basicnanoclient/rpc.py
@@ -1,0 +1,245 @@
+__package__ = "basicnanoclient"
+
+from typing import Any, Dict, Self
+import binascii
+import requests
+import random
+import base64
+import struct
+import os
+import hashlib
+import sys
+from nacl.signing import SigningKey, VerifyKey
+from nacl.encoding import RawEncoder
+from hashlib import blake2b
+
+rpc_network: str = "http://127.0.0.1:17076"
+session: requests.Session = requests.Session()
+
+class RPC():
+    """Nano RPC Client.
+
+    ```py
+    >>> from basicnanoclient.rpc import RPC
+    >>> client = RPC("http://127.0.0.1:17076")
+    >>> client.send(...)
+    ```
+    """
+
+    def __init__(self: Self, rpc_network: str) -> None:
+        """Constructor."""
+        self.rpc_network = rpc_network
+
+    def receive(
+            self: Self,
+            wallet: str,
+            account: str,
+            block: str) -> Dict[str, Any]:
+        """Receives a pending Nano block and adds it to the wallet's balance.
+
+        Args:
+            wallet (str): The Nano wallet ID.
+            account (str): The Nano account address.
+            block (str): A Nano block hash.
+
+        Returns:
+            A dictionary representing the received block.
+
+        Raises:
+            requests.exceptions.RequestException: due to the RPC request.
+        """
+        return session.post(self.rpc_network, json={
+            "action": "receive",
+            "wallet": wallet,
+            "account": account,
+            "block": block
+        }).json()
+
+    def account_info(self: Self, account: str) -> Dict[str, Any]:
+        """Retrieve information about a Nano account.
+
+        Including its balance and representative.
+
+        Args:
+            account (str): The Nano account address.
+
+        Returns:
+            A dictionary with various account information, including:
+                'frontier', 'balance', 'representative', and 'block_count'.
+
+        Raises:
+            requests.exceptions.RequestException: due to the RPC request.
+        """
+        return session.post(self.rpc_network, json={
+            "action": "account_info",
+            "representative": "true",
+            "account": account
+        }).json()
+
+    def account_key(self: Self, account: str) -> Dict[str, Any]:
+        """Retrieve the public key of a Nano account.
+
+        Args:
+            account (str): The Nano account address.
+
+        Returns:
+            A dictionary with the public key of the Nano account.
+        """
+        return session.post(self.rpc_network, json={
+            "action": "account_key",
+            "account": account
+        }).json()
+
+    def ledger(self: Self, account: str, count: int) -> Dict[str, Any]:
+        """Retrieve the transaction history for a Nano account.
+
+        Args:
+            account (str): The Nano account address.
+            count (int): The maximum number of transactions to retrieve.
+
+        Returns:
+            A dictionary containing the transaction history
+                for the Nano account.
+        """
+        return session.post(self.rpc_network, json={
+            "action": "ledger",
+            "account": account,
+            "count": count
+        }).json()
+
+    def receivable(
+            self: Self,
+            account: str,
+            count: int = 1,
+            threshold: int = 1) -> Dict[str, Any]:
+        """Retrieve a list of pending Nano transactions for an account.
+
+        Args:
+            account (str): The Nano account address.
+            count (int): The maximum number of transactions to retrieve
+                (default is 1).
+            threshold (int): The minimum amount of Nano pending in raw units
+                (default is 1 raw Nano).
+
+        Returns:
+            A dictionary containing a list of pending Nano transactions
+                for the account.
+        """
+        return session.post(self.rpc_network, json={
+            "action": "receivable",
+            "account": account,
+            "count": count,
+            "threshold": threshold,
+            "source": "true"
+        }).json()
+
+    def block_info(self: Self, block: str) -> dict:
+        """Retrieve information about a Nano block.
+
+        Args:
+            block (str): The block hash.
+
+        Returns:
+            dict: A dictionary containing information about the block.
+        """
+        return session.post(self.rpc_network, json={
+            "action": "block_info",
+            "json_block": "true",
+            "hash": block
+        }).json()
+
+    def process(self: Self, block: dict) -> dict:
+        """Process a block.
+
+        Args:
+            block (dict): The block to be processed.
+
+        Returns:
+            dict: A dictionary containing information about the block.
+        """
+        response = requests.post(self.rpc_network, json={
+            "action": "process",
+            "block": block
+        })
+        return response.json()
+
+    def sign_and_send(
+            self: Self,
+            previous: str,
+            account: str,
+            representative: str,
+            balance: str,
+            link: str,
+            key: str) -> dict:
+        """Sign and send a transaction.
+
+        Args:
+            previous (str): The previous block hash.
+            account (str): The account address.
+            representative (str): The representative address.
+            balance (str): The new account balance.
+            link (str): The link to a previous block.
+            key (str): The account private key.
+
+        Returns:
+            dict: A dictionary containing information
+                about the transaction.
+        """
+        # Create the block
+        block = self.block_create(
+            previous,
+            account,
+            representative,
+            balance,
+            link,
+            key
+        )
+
+        # Process the block
+        response = self.process(block)
+
+        return response
+
+    def send(
+            self: Self,
+            wallet: str,
+            source: str,
+            destination: str,
+            amount: int,
+            key: str) -> Dict[str, Any]:
+        """Send a specified amount of Nano from one account to another.
+
+        Args:
+            wallet (str): The Nano wallet address.
+            source (str): The Nano account address to send from.
+            destination (str): The Nano account address to send to.
+            amount (int): The amount of Nano to send in raw units.
+            key (str): The private key of the account sending the Nano.
+
+        Returns:
+            A dictionary containing information about the transaction.
+        """
+        # Retrieve the wallet info
+        wallet_info = self.wallet_info(wallet)
+        key = wallet_info.get(key)
+
+        # Retrieve the account info
+        account_info = self.account_info(source)
+        previous = account_info.get('frontier')
+        representative = account_info.get('representative')
+
+        # Calculate new balance after sending the amount
+        # Convert balance to int for calculations
+        current_balance = int(account_info.get('balance'))
+        # Subtract amount to be sent and convert back to str
+        new_balance = str(current_balance - amount)
+
+        # Sign and send the transaction
+        return self.sign_and_send(
+            previous,
+            source,
+            representative,
+            new_balance,
+            destination,
+            key
+        )
